fix: momentum_score, moat_score, fundamental_score → float (engines return decimals)
Same issue as earnings_quality_score — Pydantic rejected float values
like 29.2 when the model expected int. Changed all score fields that
come from engines to float to prevent future type mismatches.

Verified: HDFCBANK.NS now returns 200 with fair_value=735, price=795.

+++ backend/models/responses.py
@@ -53,10 +53,10 @@ class QualityOutput(BaseModel):
     earnings_quality_grade: str = ""
     earnings_quality_score: float = 0
     moat: Literal["Wide", "Narrow", "None"] = "None"
-    moat_score: int = 0
-    momentum_score: int = 0
+    moat_score: float = 0
+    momentum_score: float = 0
     momentum_grade: str = ""
-    fundamental_score: int = 0
+    fundamental_score: float = 0
     fundamental_grade: str = ""
     roe: Optional[float] = None
     de_ratio: Optional[float] = None
